Squashed commit of the following:
commit 5c8ca0004ffe13aa48e86316bfb41af789c914d5
Author: Zhang Zhuo <mycinbrin@gmail.com>
Date:   Wed Mar 29 12:26:35 2023 +0800

    change to complex selector

commit afc0f6b52daa1bb667caa107a4c2d1cab5faab25
Author: Ho Vei <noelwei@gmail.com>
Date:   Thu Feb 16 11:17:47 2023 +0800

    some trival update and fmt

commit 56b45b6d82c27472b98f4950a9f89b36e3651dbc
Author: Ho Vei <noelwei@gmail.com>
Date:   Wed Feb 15 23:33:22 2023 +0800

    upgrade spec

commit aa06d85636ba06593fadd400f546964a93e80bf8
Author: Ho Vei <noelwei@gmail.com>
Date:   Wed Feb 15 23:33:04 2023 +0800

    update spec

commit 217f40808ec5f3e9e795a597ab0f08a2efce27b6
Author: Ho Vei <noelwei@gmail.com>
Date:   Sat Apr 8 15:28:33 2023 +0800

    update hash_block_size

+++ src/poseidon/primitives.rs
@@ -475,8 +475,8 @@ impl<F: FieldExt, S: Spec<F, T, RATE>, const T: usize, const RATE: usize>
     Hash<F, S, VariableLengthIden3, T, RATE>
 {
     /// Hashes the given input.
-    pub fn hash_with_cap(mut self, message: &[F], cap: u64) -> F {
-        self.sponge.update_capacity(F::from_u128(cap as u128));
+    pub fn hash_with_cap(mut self, message: &[F], cap: u128) -> F {
+        self.sponge.update_capacity(F::from_u128(cap));
         for value in message {
             self.sponge.absorb(*value);
         }
